requirements.txt
azure-functions==1.7.0
mpmath==1.2.1
numpy==1.20.2
Pillow==8.2.0
sympy==1.7.1
